Finalise the implementation of the only-on-change feature (#35)
* Provide a custom variable to control toggling

Add the `org-appear-trigger' custom variable that controls the mode
of operation. When set to `on-change', toggle elements only when
the buffer is modified or on mouse click. When set to `always', always
toggle elements.

* Keep track of toggling

With the new `on-change' operation mode, we have to remember
if an element has been toggled previously to keep updating it
even if the user does not make new changes to the buffer.