Allow autolinks to show just brackets
Showing everything inside a link can be disorienting, as URLs tend to
be long, and showing them shifts everything by many characters.  The
URL is also usually not interesting: they're just a technical detail
of a link which is rarely inspected or updated.

This commit allows org-appear-autolinks to be set to 'just-brackets,
which instructs org-appear to show just a pair of brackets around a
link under point, and no more.  URLs can still be viewed and edited
using `org-insert-link' if needed.